+++ RELEASE_NOTES.md
@@ -1,5 +1,15 @@
 # Change Log
 
+# 1.2.3
+
+- Disabled various functional rules that were causing issues with React 18
+
+# 1.2
+
+- Moved to scoped package name
+- Added sort import rules
+- Updated dependencies
+
 # 1.1.1
 
 - Updated internals for `no-public-secrets` plugin to address plugin usage issue